Support messages and repairs larger than 2GB:  Length field in PeerChannel header is now 8 bytes.  Incremented minor protocol version from 3 to 4.  Use FilePeerMessage for large outgoing messages.
git-svn-id: svn://svn.code.sf.net/p/lockss/svn/lockss-daemon/trunk@7629 4f837ed2-42f5-46e7-a7a5-fa17313484d4

+++ test/src/org/lockss/poller/v3/FuncV3Voter.java
@@ -1,5 +1,5 @@
 /*
- * $Id: FuncV3Voter.java,v 1.24 2008-01-30 02:43:05 tlipkis Exp $
+ * $Id: FuncV3Voter.java,v 1.25 2008-11-02 21:13:48 tlipkis Exp $
  */
 
 /*
@@ -378,7 +378,7 @@ public class FuncV3Voter extends LockssTestCase {
     public PeerMessage newPeerMessage() {
       throw new UnsupportedOperationException("Not implemented");
     }
-    public PeerMessage newPeerMessage(int estSize) {
+    public PeerMessage newPeerMessage(long estSize) {
       throw new UnsupportedOperationException("Not implemented");
     }
     public void registerMessageHandler(int protocol, MessageHandler handler) {
